requirements.txt
google-genai>=1.0.0
playwright>=1.44.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
aiosqlite>=0.20.0
pillow>=10.0.0
aiofiles>=23.0.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
sse-starlette>=2.0.0
